+++ R/dct.R
@@ -296,7 +296,7 @@ reframe_with_dct <- function(
   check_grouping(.data, {{ .by }})
 
 
-  if (quo_is_null(enquo(.time_col))) {
+  if (quo_is_null(enquo(.time_col)) & options::opt("tidynorm.warnings")) {
     cli_par()
     cli_inform(
       c(
@@ -434,7 +434,7 @@ reframe_with_idct <- function(
   # make sure groupings are ok
   check_grouping(.data, {{ .by }})
 
-  if (quo_is_null(cols$.param_col)) {
+  if (quo_is_null(cols$.param_col) & options::opt("tidynorm.warnings")) {
     cli_par()
     cli_inform(
       c(
@@ -504,7 +504,7 @@ reframe_with_idct <- function(
     by = unique(joining)
   ) |>
     relocate(
-      !!targets,
+      tidyselect::starts_with(names(target_pos)),
       .before = min(target_pos)
     )
   return(out_df)
@@ -623,6 +623,7 @@ reframe_with_dct_smooth <- function(
     !!targets,
     .token_id_col = {{ .token_id_col }},
     .by = !!by_grouping,
+    .n = !!sym(".n"),
     .param_col = !!sym(".param"),
     .rate = .rate,
     .accel = .accel