+++ R/dct_norm.R
@@ -3,6 +3,8 @@
 #' @inheritParams norm_track_generic
 #' @param .data A data frame of formant DCT coefficients
 #' @param .param_col A column identifying the DCT parameter number.
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #' @details
 #'
 #' The following `norm_dct_*` procedures were built on top of
@@ -80,7 +82,7 @@ norm_dct_generic <- function(
     .by_formant = FALSE,
     .by_token = FALSE,
     .names = "{.formant}_n",
-    .silent = FALSE,
+    .silent = opt("tidynorm.silent"),
     .drop_orig = FALSE,
     .call = caller_env()) {
   if (env_name(.call) == "global") {
@@ -292,11 +294,7 @@ norm_dct_generic <- function(
     norm_info
   )
 
-  if (!.silent) {
-    wrap_up(
-      normed
-    )
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -305,6 +303,8 @@ norm_dct_generic <- function(
 #'
 #' @inheritParams norm_dct_generic
 #' @inheritParams norm_track_lobanov
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #'
@@ -341,7 +341,7 @@ norm_dct_lobanov <- function(
     .by = NULL,
     .param_col = NULL,
     .names = "{.formant}_z",
-    .silent = FALSE,
+    .silent = opt("tidynorm.silent"),
     .drop_orig = FALSE) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
@@ -368,9 +368,7 @@ norm_dct_lobanov <- function(
     list(.norm_procedure = "tidynorm::norm_dct_lobanov")
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -381,6 +379,8 @@ norm_dct_lobanov <- function(
 #'
 #' @inheritParams norm_track_generic
 #' @inheritParams norm_dct_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #'
@@ -428,7 +428,7 @@ norm_dct_nearey <- function(
     .param_col = NULL,
     .drop_orig = FALSE,
     .names = "{.formant}_lm",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -456,9 +456,7 @@ norm_dct_nearey <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -468,6 +466,8 @@ norm_dct_nearey <- function(
 #'
 #' @inheritParams norm_track_generic
 #' @inheritParams norm_dct_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #'
 #' @details
@@ -502,7 +502,7 @@ norm_dct_deltaF <- function(
     .param_col = NULL,
     .drop_orig = FALSE,
     .names = "{.formant}_df",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -528,9 +528,7 @@ norm_dct_deltaF <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -539,6 +537,8 @@ norm_dct_deltaF <- function(
 #'
 #' @inheritParams norm_track_generic
 #' @inheritParams norm_dct_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #' This is a modified version of the Watt & Fabricius Method. The original
@@ -577,7 +577,7 @@ norm_dct_wattfab <- function(
     .param_col = NULL,
     .drop_orig = FALSE,
     .names = "{.formant}_wf",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -604,9 +604,7 @@ norm_dct_wattfab <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -614,6 +612,8 @@ norm_dct_wattfab <- function(
 #' Bark Difference DCT Normalization
 #' @inheritParams norm_track_generic
 #' @inheritParams norm_dct_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #' @details
 #'
 #' **Important**: This function assumes that the DCT
@@ -656,7 +656,7 @@ norm_dct_barkz <- function(
     .param_col = NULL,
     .drop_orig = FALSE,
     .names = "{.formant}_bz",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -698,9 +698,7 @@ norm_dct_barkz <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }