+++ R/track_norm.R
@@ -31,6 +31,8 @@
 #' should be returned.
 #' @param .silent Whether or not the informational message should be printed.
 #' @param .call Used for internal purposes.
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #' The following `norm_track_*` procedures were built on top of
@@ -122,7 +124,7 @@ norm_track_generic <- function(
     .return_dct = FALSE,
     .drop_orig = FALSE,
     .names = "{.formant}_n",
-    .silent = FALSE,
+    .silent = opt("tidynorm.silent"),
     .call = caller_env()) {
   if (env_name(.call) == "global") {
     .call2 <- current_env()
@@ -340,9 +342,7 @@ norm_track_generic <- function(
     norm_info
   )
 
-  if (!.silent) {
-    wrap_up(normed_track)
-  }
+  wrap_up(normed_track, .silent)
 
   return(normed_track)
 }
@@ -350,6 +350,8 @@ norm_track_generic <- function(
 #' Lobanov Track Normalization
 #'
 #' @inheritParams norm_track_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #'
@@ -389,7 +391,7 @@ norm_track_lobanov <- function(
     .return_dct = FALSE,
     .drop_orig = FALSE,
     .names = "{.formant}_z",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -420,9 +422,7 @@ norm_track_lobanov <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -430,6 +430,8 @@ norm_track_lobanov <- function(
 #' Nearey Track Normalization
 #'
 #' @inheritParams norm_track_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #' When formant extrinsic:
@@ -474,7 +476,7 @@ norm_track_nearey <- function(
     .return_dct = FALSE,
     .drop_orig = FALSE,
     .names = "{.formant}_lm",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -505,9 +507,7 @@ norm_track_nearey <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -517,6 +517,8 @@ norm_track_nearey <- function(
 #' Delta F Track Normalization
 #'
 #' @inheritParams norm_track_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #'
 #' @details
@@ -553,7 +555,7 @@ norm_track_deltaF <- function(
     .return_dct = FALSE,
     .drop_orig = FALSE,
     .names = "{.formant}_df",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -583,9 +585,7 @@ norm_track_deltaF <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
@@ -594,6 +594,8 @@ norm_track_deltaF <- function(
 #' Watt and Fabricius Track normalization
 #'
 #' @inheritParams norm_track_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
 #'
 #' @details
 #' This is a modified version of the Watt & Fabricius Method. The original
@@ -634,7 +636,7 @@ norm_track_wattfab <- function(
     .return_dct = FALSE,
     .drop_orig = FALSE,
     .names = "{.formant}_wf",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -665,15 +667,16 @@ norm_track_wattfab <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }
 
 #' Bark Difference Track Normalization
 #' @inheritParams norm_track_generic
+#' @eval options::as_params(".silent" = "tidynorm.silent")
+#' @importFrom options opt
+#'
 #' @details
 #' This is a within-token normalization technique. First all formants are
 #' converted to Bark (see [hz_to_bark]), then, within each token, F3 is
@@ -713,7 +716,7 @@ norm_track_barkz <- function(
     .return_dct = FALSE,
     .drop_orig = FALSE,
     .names = "{.formant}_bz",
-    .silent = FALSE) {
+    .silent = opt("tidynorm.silent")) {
   args <- names(call_match())
   fmls <- names(fn_fmls())
   check_args(args, fmls)
@@ -761,9 +764,7 @@ norm_track_barkz <- function(
     )
   )
 
-  if (!.silent) {
-    wrap_up(normed)
-  }
+  wrap_up(normed, .silent)
 
   return(normed)
 }